Undo fetch refactor, remove stray console.logs, restore child clone PP
